Change Modbus prefix
Rename Modbus files and functions to use new prefix.

The modbus-* prefix is already used by libmodbus, which leads to
confusion. The new prefix is iface-modbus, which offers better
consistency with future interfaces and offers a clear distinction
from libmodbus's.
