+++ Makefile.am
@@ -19,9 +19,9 @@ src_thingd_SOURCES = src/main.c \
 			src/cloud.c src/cloud.h \
 			src/device.c src/device.h \
 			src/storage.c src/storage.h \
-			src/modbus-interface.c src/modbus-interface.h \
-			src/modbus-tcp.c src/modbus-rtu.c \
-			src/modbus-driver.h \
+			src/iface-modbus.c src/iface-modbus.h \
+			src/iface-modbus-tcp.c src/iface-modbus-rtu.c \
+			src/iface-modbus-driver.h \
 			src/settings.c src/settings.h \
 			src/knot-config.c src/knot-config.h \
 			src/poll.c src/poll.h
