+++ src/iface-modbus.h
@@ -17,12 +17,12 @@
 extern struct modbus_driver tcp;
 extern struct modbus_driver rtu;
 
-typedef void (*modbus_connected_cb_t) (void *user_data);
-typedef void (*modbus_disconnected_cb_t) (void *user_data);
+typedef void (*iface_modbus_connected_cb_t) (void *user_data);
+typedef void (*iface_modbus_disconnected_cb_t) (void *user_data);
 
-int modbus_read_data(int reg_addr, int bit_offset, knot_value_type *out);
-int modbus_start(const char *url, int slave_id,
-		 modbus_connected_cb_t connected_cb,
-		 modbus_disconnected_cb_t disconnected_cb,
-		 void *user_data);
-void modbus_stop(void);
+int iface_modbus_read_data(int reg_addr, int bit_offset, knot_value_type *out);
+int iface_modbus_start(const char *url, int slave_id,
+		       iface_modbus_connected_cb_t connected_cb,
+		       iface_modbus_disconnected_cb_t disconnected_cb,
+		       void *user_data);
+void iface_modbus_stop(void);
